Added spotifylistener module that has a lambda-handler for the action of archiving tracks heard in the last hour
The function queries Spotify Web API with an access-token to get the
most recenntly heard tracks in the last hour.

It extracts when the track was played, its id, duration and calculates
how long it was listened to.

This information is then uploaded to S3 as a CSV file with the name
reflecting the beginning of the hour.