Added SAM template for SoundprintSpotifyTracker lambda with a barebones handler
Added spotifytracker.py to act as lambda-handler for lambda function
that is to be triggered upon creation of S3 file by spotifylistener. The
function implementation of spotifytracker is empty for now.

Also added SAM template to create the lambda with triggers for S3 Events
upon creation of csv files under the history/listening/ directory.