Modified spotifytracker and spotifyarchiver to encode lists into strings before inserting in dataframe
In spotifytracker, we pull the list of artists for a track from Spotify.
With this change, this list of artist-ids is encoded into a single
string representing this entire list and that string is stored in the
dataframe and uploaded to S3. Using a codec makes it easier for readers
of this archive to parse the list of artists from the field.

Similarly, spotifyalbumer pulls the list of genres for an album. This
change adds a similar codec for the list of genres to the archived file
format.