Created SNS Topic for notifcations on creation of history/tracks S3 files
A single S3 event cannot be used to trigger multiple Lambda functions.

Since we need to trigger both SoundprintSpotifyAlbumer and
SoundprintSpotifyArtister lambdas upon creation of CSV files under
history/tracks/ in the S3 bucket, we have to do so via an SNS topic
which can then be used to trigger these lambdas.

This change creates a new SNS topic and modifies the S3 bucket with
notification configuration to publish notifications to the SNS topic
upon creation of .csv objects under histroy/tracks/

SoundprintSpotifyAlbumer and SoundprintSpotifyArtister then use these
SNS notifications to be triggered.