Added implementation logic for spotifyrdsarchiver
Lambda handler for the action of taking the recently generated Spotify-history data and archiving the records
into Aurora DB. Gets triggered after joining multiple types of metadata is complete and the joined file name
is provided.
If the Aurora table does not exist, it creates the table. If there are records to be inserted, the Aurora serverless
cluster is first woken up with exponential backoff until it is ready to receive SQL requests.